Modificaciones

+++ Entrega 2/UPMFit22_09/src/main/java/clases/InterfazAdministrador.java
@@ -9,7 +9,7 @@ public interface InterfazAdministrador {
 
 	public void altaMonitor();
 
-	public double getTelefono();
+	public int getTelefono();
 
 	/**
 	 *
